Improve EUI handling of NCN changes
This updates the EUI NCN handling logic as follows:

- If the document has an NCN of any kind and the field is blanked, it will reject the change.
- If the NCN field is present, and *either* it doesn't already have one *or* its value has changed, it'll try process a change.
- Submitted NCN is now cast to a proper `NeutralCitationString` so it's subject to validation flows.

+++ judgments/views/judgment_edit.py
@@ -6,6 +6,7 @@ from django.contrib import messages
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.views.generic import View
+from ds_caselaw_utils.types import NeutralCitationString
 
 from judgments.utils import (
     MoveJudgmentError,
@@ -44,49 +45,65 @@ class EditJudgmentView(View):
         # TODO: confirm that the stub we are about to set does not already have a mapping that isn't for this document
 
         try:
-            # confirm that the identifier isn't used elsewhere
-            new_citation = request.POST["neutral_citation"]
-            verify_stub_not_used(judgment_uri, new_citation)
+            form_neutral_citation = request.POST["neutral_citation"]
 
-            # Set name
-            new_name = request.POST["metadata_name"]
-            api_client.set_document_name(judgment_uri, new_name)
+            existing_preferred_ncn = judgment.identifiers.preferred(NeutralCitationNumber)
 
-            # Set neutral citation within the judgment text
-            new_citation = request.POST["neutral_citation"]
-            api_client.set_judgment_citation(judgment_uri, new_citation)
+            # If the NCN is now empty and wasn't before, fail out and don't do anything else.
+            # TODO: This is horrible nested logic and needs refactoring into something with better guard clause behaviours
 
-            # Set neutral citation in the identifiers
-            # Note that this currently does not preserve the UUID if the
-            # neutral citation is unchanged.
-
-            judgment.identifiers.delete_type(NeutralCitationNumber)
-            judgment.identifiers.add(NeutralCitationNumber(value=new_citation))
-            judgment.save_identifiers()
-
-            # Set court
-            new_court = request.POST["court"]
-            api_client.set_document_court_and_jurisdiction(judgment_uri, new_court)
-
-            # Date
-            new_date = request.POST["judgment_date"]
-            try:
-                new_date_as_date = datetime.datetime.strptime(new_date, r"%d %b %Y")
-            except ValueError:
+            if form_neutral_citation == "" and existing_preferred_ncn:
                 messages.error(
                     request,
-                    f"Could not parse the date '{new_date}', should be like '2 Jan 2024'.",
+                    "A document with a pre-existing NCN cannot be saved with an empty NCN",
                 )
-                return HttpResponseRedirect(
-                    reverse(
-                        "full-text-html",
-                        kwargs={"document_uri": kwargs["document_uri"]},
-                    ),
-                )
-            new_date_as_iso = new_date_as_date.strftime(r"%Y-%m-%d")
-            api_client.set_judgment_date(judgment_uri, new_date_as_iso)
 
-            messages.success(request, "Document successfully updated")
+            else:
+                # If there is a submitted NCN, and EITHER one doesn't already exist OR one does exist and the value is different, process changes
+                if form_neutral_citation != "" and (
+                    existing_preferred_ncn is None or form_neutral_citation != existing_preferred_ncn.value
+                ):
+                    # Convert the submitted NCN to a proper class (which will perform validation)
+                    neutral_citation = NeutralCitationString(form_neutral_citation)
+
+                    # Confirm that the identifier isn't used elsewhere
+                    verify_stub_not_used(judgment_uri, neutral_citation)
+
+                    # Set neutral citation in the identifiers
+                    judgment.identifiers.delete_type(NeutralCitationNumber)
+                    judgment.identifiers.add(NeutralCitationNumber(value=neutral_citation))
+                    judgment.save_identifiers()
+
+                    # Set neutral citation in the document XML
+                    api_client.set_judgment_citation(judgment_uri, neutral_citation)
+
+                # Set name
+                new_name = request.POST["metadata_name"]
+                api_client.set_document_name(judgment_uri, new_name)
+
+                # Set court
+                new_court = request.POST["court"]
+                api_client.set_document_court_and_jurisdiction(judgment_uri, new_court)
+
+                # Date
+                new_date = request.POST["judgment_date"]
+                try:
+                    new_date_as_date = datetime.datetime.strptime(new_date, r"%d %b %Y")
+                except ValueError:
+                    messages.error(
+                        request,
+                        f"Could not parse the date '{new_date}', should be like '2 Jan 2024'.",
+                    )
+                    return HttpResponseRedirect(
+                        reverse(
+                            "full-text-html",
+                            kwargs={"document_uri": kwargs["document_uri"]},
+                        ),
+                    )
+                new_date_as_iso = new_date_as_date.strftime(r"%Y-%m-%d")
+                api_client.set_judgment_date(judgment_uri, new_date_as_iso)
+
+                messages.success(request, "Document successfully updated")
 
         except (MoveJudgmentError, NeutralCitationToUriError, StubAlreadyUsedError) as e:
             messages.error(
